Chore: move fetchBRbyLoc to appropriate component
