+++ src/Components/search/Search.js
@@ -1,21 +1,21 @@
 import './Search.css';
-// import { useEffect } from 'react';
-// import { fetchBRsByLocNoParams } from '../../apiCalls';
+import { useEffect } from 'react';
+import { fetchBRsByLocNoParams } from '../../apiCalls';
 
 const Search = () => {
-    // const lat = 39.753604
-    // const lon = -104.428580
-    // useEffect(() => {
-    //     fetch()
-    //   }, [])
-    //  async function fetch() {
-    //   try {
-    //     const tryFetch = await fetchBRsByLocNoParams(lat, lon)
-    //     console.log(tryFetch)
-    //   } catch(error) {
-    //     throw error
-    //   }
-    //  }
+    const lat = 39.753604
+    const lon = -104.428580
+    useEffect(() => {
+        fetch()
+      }, [])
+     async function fetch() {
+      try {
+        const tryFetch = await fetchBRsByLocNoParams(lat, lon)
+        console.log(tryFetch)
+      } catch(error) {
+        throw error
+      }
+     }
     return (
         <>
         </>
